+++ paper/src/main/java/com/badbones69/crazycrates/CrazyCrates.java
@@ -23,11 +23,13 @@ import com.badbones69.crazycrates.support.placeholders.PlaceholderAPISupport;
 import com.badbones69.crazycrates.tasks.BukkitUserManager;
 import com.badbones69.crazycrates.tasks.InventoryManager;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
-import com.ryderbelserion.vital.paper.VitalPaper;
 import com.ryderbelserion.vital.paper.enums.Support;
 import com.ryderbelserion.vital.paper.files.config.FileManager;
+import me.arcaniax.hdb.api.HeadDatabaseAPI;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
 import java.util.List;
 import java.util.Timer;
 import static com.badbones69.crazycrates.api.utils.MiscUtils.registerPermissions;
@@ -45,6 +47,7 @@ public class CrazyCrates extends JavaPlugin {
     private BukkitUserManager userManager;
     private CrateManager crateManager;
     private FileManager fileManager;
+    private HeadDatabaseAPI api;
 
     private Server instance;
 
@@ -61,6 +64,10 @@ public class CrazyCrates extends JavaPlugin {
 
         // Register permissions that we need.
         registerPermissions();
+
+        if (Support.head_database.isEnabled()) {
+            this.api = new HeadDatabaseAPI();
+        }
 
         this.inventoryManager = new InventoryManager();
         this.crateManager = new CrateManager();
@@ -162,6 +169,14 @@ public class CrazyCrates extends JavaPlugin {
 
     public @NotNull final FileManager getFileManager() {
         return this.fileManager;
+    }
+
+    public @Nullable final HeadDatabaseAPI getApi() {
+        if (this.api == null) {
+            return null;
+        }
+
+        return this.api;
     }
 
     public @NotNull final Server getInstance() {